requirements.txt
colorama==0.4.3
deep_translator==1.10.1
insta_scrape==2.1.2
#kapture==1.1.8
matplotlib==3.4.3
nltools==0.4.7
numpy==1.21.5
opencv_python==4.4.0.46
pandas==1.2.4
Pillow==9.5.0
pytangle==0.0.2
python_dateutil==2.8.2
Requests==2.30.0
scikit_learn==0.23.1
seaborn==0.11.2
spacy==3.2.1
spacy_langdetect==0.1.2
swifter==1.3.4
#torch==1.7.1+cu101
#torchvision==0.8.2+cu101
tqdm==4.62.2
